Split RSA256 top module
